+++ liquidity/liquidity.go
@@ -124,6 +124,7 @@ var (
 		AutoFeeBudget:              defaultBudget,
 		MaxAutoInFlight:            defaultMaxInFlight,
 		ChannelRules:               make(map[lnwire.ShortChannelID]*ThresholdRule),
+		PeerRules:                  make(map[route.Vertex]*ThresholdRule),
 		FailureBackOff:             defaultFailureBackoff,
 		SweepFeeRateLimit:          defaultSweepFeeRateLimit,
 		SweepConfTarget:            loop.DefaultSweepConfTarget,
@@ -181,6 +182,11 @@ var (
 
 	// ErrNoRules is returned when no rules are set for swap suggestions.
 	ErrNoRules = errors.New("no rules set for autoloop")
+
+	// ErrExclusiveRules is returned when a set of rules that may not be
+	// set together are specified.
+	ErrExclusiveRules = errors.New("channel and peer rules must be " +
+		"exclusive")
 )
 
 // Config contains the external functionality required to run the
@@ -289,27 +295,41 @@ type Parameters struct {
 	ClientRestrictions Restrictions
 
 	// ChannelRules maps a short channel ID to a rule that describes how we
-	// would like liquidity to be managed.
+	// would like liquidity to be managed. These rules and PeerRules are
+	// exclusively set to prevent overlap between peer and channel rules.
 	ChannelRules map[lnwire.ShortChannelID]*ThresholdRule
+
+	// PeerRules maps a peer's pubkey to a rule that applies to all the
+	// channels that we have with the peer collectively. These rules and
+	// ChannelRules are exclusively set to prevent overlap between peer
+	// and channel rules map to avoid ambiguity.
+	PeerRules map[route.Vertex]*ThresholdRule
 }
 
 // String returns the string representation of our parameters.
 func (p Parameters) String() string {
-	channelRules := make([]string, 0, len(p.ChannelRules))
+	ruleList := make([]string, 0, len(p.ChannelRules)+len(p.PeerRules))
 
 	for channel, rule := range p.ChannelRules {
-		channelRules = append(
-			channelRules, fmt.Sprintf("%v: %v", channel, rule),
+		ruleList = append(
+			ruleList, fmt.Sprintf("Channel: %v: %v", channel, rule),
 		)
 	}
 
-	return fmt.Sprintf("channel rules: %v, failure backoff: %v, sweep "+
+	for peer, rule := range p.PeerRules {
+		ruleList = append(
+			ruleList, fmt.Sprintf("Peer: %v: %v", peer, rule),
+		)
+
+	}
+
+	return fmt.Sprintf("rules: %v, failure backoff: %v, sweep "+
 		"fee rate limit: %v, sweep conf target: %v, maximum prepay: "+
 		"%v, maximum miner fee: %v, maximum swap fee ppm: %v, maximum "+
 		"routing fee ppm: %v, maximum prepay routing fee ppm: %v, "+
 		"auto budget: %v, budget start: %v, max auto in flight: %v, "+
 		"minimum swap size=%v, maximum swap size=%v",
-		strings.Join(channelRules, ","), p.FailureBackOff,
+		strings.Join(ruleList, ","), p.FailureBackOff,
 		p.SweepFeeRateLimit, p.SweepConfTarget, p.MaximumPrepay,
 		p.MaximumMinerFee, p.MaximumSwapFeePPM,
 		p.MaximumRoutingFeePPM, p.MaximumPrepayRoutingFeePPM,
@@ -317,9 +337,54 @@ func (p Parameters) String() string {
 		p.ClientRestrictions.Minimum, p.ClientRestrictions.Maximum)
 }
 
-// validate checks whether a set of parameters is valid. It takes the minimum
-// confirmations we allow for sweep confirmation target as a parameter.
-func (p Parameters) validate(minConfs int32, server *Restrictions) error {
+// haveRules returns a boolean indicating whether we have any rules configured.
+func (p Parameters) haveRules() bool {
+	if len(p.ChannelRules) != 0 {
+		return true
+	}
+
+	if len(p.PeerRules) != 0 {
+		return true
+	}
+
+	return false
+}
+
+// validate checks whether a set of parameters is valid. Our set of currently
+// open channels are required to check that there is no overlap between the
+// rules set on a per-peer level, and those set for specific channels. We can't
+// allow both, because then we're trying to cater for two separate liquidity
+// goals on the same channel. Since we use short channel ID, we don't need to
+// worry about pending channels (users would need to work very hard to get the
+// short channel ID for a pending channel). Likewise, we don't care about closed
+// channels, since there is no action that may occur on them, and we want to
+// allow peer-level rules to be set once a channel which had a specific rule
+// has been closed. It takes the minimum confirmations we allow for sweep
+// confirmation target as a parameter.
+// TODO(carla): prune channels that have been closed from rules.
+func (p Parameters) validate(minConfs int32, openChans []lndclient.ChannelInfo,
+	server *Restrictions) error {
+
+	// First, we check that the rules on a per peer and per channel do not
+	// overlap, since this could lead to contractions.
+	for _, channel := range openChans {
+		// If we don't have a rule for the peer, there's no way we have
+		// an overlap between this peer and the channel.
+		_, ok := p.PeerRules[channel.PubKeyBytes]
+		if !ok {
+			continue
+		}
+
+		shortID := lnwire.NewShortChanIDFromInt(channel.ChannelID)
+		_, ok = p.ChannelRules[shortID]
+		if ok {
+			log.Debugf("Rules for peer: %v and its channel: %v "+
+				"can't both be set", channel.PubKeyBytes, shortID)
+
+			return ErrExclusiveRules
+		}
+	}
+
 	for channel, rule := range p.ChannelRules {
 		if channel.ToUint64() == 0 {
 			return ErrZeroChannelID
@@ -328,6 +393,13 @@ func (p Parameters) validate(minConfs int32, server *Restrictions) error {
 		if err := rule.validate(); err != nil {
 			return fmt.Errorf("channel: %v has invalid rule: %v",
 				channel.ToUint64(), err)
+		}
+	}
+
+	for peer, rule := range p.PeerRules {
+		if err := rule.validate(); err != nil {
+			return fmt.Errorf("peer: %v has invalid rule: %v",
+				peer, err)
 		}
 	}
 
@@ -483,7 +555,12 @@ func (m *Manager) SetParameters(ctx context.Context, params Parameters) error {
 		return err
 	}
 
-	err = params.validate(m.cfg.MinimumConfirmations, restrictions)
+	channels, err := m.cfg.Lnd.Client.ListChannels(ctx)
+	if err != nil {
+		return err
+	}
+
+	err = params.validate(m.cfg.MinimumConfirmations, channels, restrictions)
 	if err != nil {
 		return err
 	}
@@ -508,6 +585,16 @@ func cloneParameters(params Parameters) Parameters {
 	for channel, rule := range params.ChannelRules {
 		ruleCopy := *rule
 		paramCopy.ChannelRules[channel] = &ruleCopy
+	}
+
+	paramCopy.PeerRules = make(
+		map[route.Vertex]*ThresholdRule,
+		len(params.PeerRules),
+	)
+
+	for peer, rule := range params.PeerRules {
+		ruleCopy := *rule
+		paramCopy.PeerRules[peer] = &ruleCopy
 	}
 
 	return paramCopy
@@ -566,12 +653,28 @@ type Suggestions struct {
 	// DisqualifiedChans maps the set of channels that we do not recommend
 	// swaps on to the reason that we did not recommend a swap.
 	DisqualifiedChans map[lnwire.ShortChannelID]Reason
+
+	// Disqualified peers maps the set of peers that we do not recommend
+	// swaps for to the reason that they were excluded.
+	DisqualifiedPeers map[route.Vertex]Reason
 }
 
 func newSuggestions() *Suggestions {
 	return &Suggestions{
 		DisqualifiedChans: make(map[lnwire.ShortChannelID]Reason),
+		DisqualifiedPeers: make(map[route.Vertex]Reason),
 	}
+}
+
+func (s *Suggestions) addSwap(swap swapSuggestion) error {
+	out, ok := swap.(*loopOutSwapSuggestion)
+	if !ok {
+		return fmt.Errorf("unexpected swap type: %T", swap)
+	}
+
+	s.OutSwaps = append(s.OutSwaps, out.OutRequest)
+
+	return nil
 }
 
 // singleReasonSuggestion is a helper function which returns a set of
@@ -582,6 +685,10 @@ func (m *Manager) singleReasonSuggestion(reason Reason) *Suggestions {
 
 	for id := range m.params.ChannelRules {
 		resp.DisqualifiedChans[id] = reason
+	}
+
+	for peer := range m.params.PeerRules {
+		resp.DisqualifiedPeers[peer] = reason
 	}
 
 	return resp
@@ -601,7 +708,7 @@ func (m *Manager) SuggestSwaps(ctx context.Context, autoloop bool) (
 
 	// If we have no rules set, exit early to avoid unnecessary calls to
 	// lnd and the server.
-	if len(m.params.ChannelRules) == 0 {
+	if !m.params.haveRules() {
 		return nil, ErrNoRules
 	}
 
@@ -688,76 +795,78 @@ func (m *Manager) SuggestSwaps(ctx context.Context, autoloop bool) (
 		return nil, err
 	}
 
+	peerChannels := make(map[route.Vertex]*balances)
+	for _, channel := range channels {
+		bal, ok := peerChannels[channel.PubKeyBytes]
+		if !ok {
+			bal = &balances{}
+		}
+
+		chanID := lnwire.NewShortChanIDFromInt(channel.ChannelID)
+		bal.channels = append(bal.channels, chanID)
+		bal.capacity += channel.Capacity
+		bal.incoming += channel.RemoteBalance
+		bal.outgoing += channel.LocalBalance
+		bal.pubkey = channel.PubKeyBytes
+
+		peerChannels[channel.PubKeyBytes] = bal
+	}
+
 	// Get a summary of the channels and peers that are not eligible due
 	// to ongoing swaps.
 	traffic := m.currentSwapTraffic(loopOut, loopIn)
 
 	var (
-		suggestions  []loop.OutRequest
-		disqualified = make(map[lnwire.ShortChannelID]Reason)
+		suggestions []swapSuggestion
+		resp        = newSuggestions()
 	)
+
+	for peer, balances := range peerChannels {
+		rule, haveRule := m.params.PeerRules[peer]
+		if !haveRule {
+			continue
+		}
+
+		suggestion, err := m.suggestSwap(
+			ctx, traffic, balances, rule, restrictions, autoloop,
+		)
+		var reasonErr *reasonError
+		if errors.As(err, &reasonErr) {
+			resp.DisqualifiedPeers[peer] = reasonErr.reason
+			continue
+		}
+
+		if err != nil {
+			return nil, err
+		}
+
+		suggestions = append(suggestions, suggestion)
+	}
 
 	for _, channel := range channels {
 		balance := newBalances(channel)
 
-		rule, ok := m.params.ChannelRules[balance.channelID]
+		channelID := lnwire.NewShortChanIDFromInt(channel.ChannelID)
+		rule, ok := m.params.ChannelRules[channelID]
 		if !ok {
 			continue
 		}
 
-		// Check whether we can perform a swap, adding the channel to
-		// our set of disqualified swaps if it is not eligible.
-		reason := traffic.maySwap(channel.PubKeyBytes, balance.channelID)
-		if reason != ReasonNone {
-			disqualified[balance.channelID] = reason
-			continue
-		}
-
-		// We can have nil suggestions in the case where no action is
-		// required, so we skip over them.
-		suggestion := rule.suggestSwap(balance, restrictions)
-		if suggestion == nil {
-			disqualified[balance.channelID] = ReasonLiquidityOk
-			continue
-		}
-
-		// Get a quote for a swap of this amount.
-		quote, err := m.cfg.LoopOutQuote(
-			ctx, &loop.LoopOutQuoteRequest{
-				Amount:                  suggestion.Amount,
-				SweepConfTarget:         m.params.SweepConfTarget,
-				SwapPublicationDeadline: m.cfg.Clock.Now(),
-			},
+		suggestion, err := m.suggestSwap(
+			ctx, traffic, balance, rule, restrictions, autoloop,
 		)
+
+		var reasonErr *reasonError
+		if errors.As(err, &reasonErr) {
+			resp.DisqualifiedChans[channelID] = reasonErr.reason
+			continue
+		}
+
 		if err != nil {
 			return nil, err
 		}
 
-		log.Debugf("quote for suggestion: %v, swap fee: %v, "+
-			"miner fee: %v, prepay: %v", suggestion, quote.SwapFee,
-			quote.MinerFee, quote.PrepayAmount)
-
-		// Check that the estimated fees for the suggested swap are
-		// below the fee limits configured by the manager.
-		feeReason := m.checkFeeLimits(quote, suggestion.Amount)
-		if feeReason != ReasonNone {
-			disqualified[balance.channelID] = feeReason
-			continue
-		}
-
-		outRequest, err := m.makeLoopOutRequest(
-			ctx, suggestion, quote, autoloop,
-		)
-		if err != nil {
-			return nil, err
-		}
-		suggestions = append(suggestions, outRequest)
-	}
-
-	// Finally, run through all possible swaps, excluding swaps that are
-	// not feasible due to fee or budget restrictions.
-	resp := &Suggestions{
-		DisqualifiedChans: disqualified,
+		suggestions = append(suggestions, suggestion)
 	}
 
 	// If we have no swaps to execute after we have applied all of our
@@ -768,7 +877,7 @@ func (m *Manager) SuggestSwaps(ctx context.Context, autoloop bool) (
 
 	// Sort suggestions by amount in descending order.
 	sort.SliceStable(suggestions, func(i, j int) bool {
-		return suggestions[i].Amount > suggestions[j].Amount
+		return suggestions[i].amount() > suggestions[j].amount()
 	})
 
 	// Run through our suggested swaps in descending order of amount and
@@ -777,11 +886,14 @@ func (m *Manager) SuggestSwaps(ctx context.Context, autoloop bool) (
 
 	// setReason is a helper that adds a swap's channels to our disqualified
 	// list with the reason provided.
-	setReason := func(reason Reason, swap loop.OutRequest) {
-		for _, id := range swap.OutgoingChanSet {
-			chanID := lnwire.NewShortChanIDFromInt(id)
+	setReason := func(reason Reason, swap swapSuggestion) {
+		for _, channel := range swap.channels() {
+			_, ok := m.params.ChannelRules[channel]
+			if !ok {
+				continue
+			}
 
-			resp.DisqualifiedChans[chanID] = reason
+			resp.DisqualifiedChans[channel] = reason
 		}
 	}
 
@@ -805,23 +917,91 @@ func (m *Manager) SuggestSwaps(ctx context.Context, autoloop bool) (
 			continue
 		}
 
-		fees := worstCaseOutFees(
-			swap.MaxPrepayRoutingFee, swap.MaxSwapRoutingFee,
-			swap.MaxSwapFee, swap.MaxMinerFee, swap.MaxPrepayAmount,
-		)
+		fees := swap.fees()
 
 		// If the maximum fee we expect our swap to use is less than the
 		// amount we have available, we add it to our set of swaps that
 		// fall within the budget and decrement our available amount.
 		if fees <= available {
 			available -= fees
-			resp.OutSwaps = append(resp.OutSwaps, swap)
+
+			if err := resp.addSwap(swap); err != nil {
+				return nil, err
+			}
 		} else {
 			setReason(ReasonBudgetInsufficient, swap)
 		}
 	}
 
 	return resp, nil
+}
+
+// suggestSwap checks whether we can currently perform a swap, and creates a
+// swap request for the rule provided.
+func (m *Manager) suggestSwap(ctx context.Context, traffic *swapTraffic,
+	balance *balances, rule *ThresholdRule, restrictions *Restrictions,
+	autoloop bool) (swapSuggestion, error) {
+
+	// Check whether we can perform a swap.
+	err := traffic.maySwap(balance.pubkey, balance.channels)
+	if err != nil {
+		return nil, err
+	}
+
+	// We can have nil suggestions in the case where no action is
+	// required, so we skip over them.
+	amount := rule.swapAmount(balance, restrictions)
+	if amount == 0 {
+		return nil, newReasonError(ReasonLiquidityOk)
+	}
+
+	swap, err := m.loopOutSwap(ctx, amount, balance, autoloop)
+	if err != nil {
+		return nil, err
+	}
+
+	return &loopOutSwapSuggestion{
+		OutRequest: *swap,
+	}, nil
+}
+
+// loopOutSwap creates a loop out swap with the amount provided for the balance
+// described by the balance set provided. A reason that indicates whether we
+// can swap is returned. If this value is not ReasonNone, there is no possible
+// swap and the loop out request returned will be nil.
+func (m *Manager) loopOutSwap(ctx context.Context, amount btcutil.Amount,
+	balance *balances, autoloop bool) (*loop.OutRequest, error) {
+
+	quote, err := m.cfg.LoopOutQuote(
+		ctx, &loop.LoopOutQuoteRequest{
+			Amount:                  amount,
+			SweepConfTarget:         m.params.SweepConfTarget,
+			SwapPublicationDeadline: m.cfg.Clock.Now(),
+		},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	log.Debugf("quote for suggestion: %v, swap fee: %v, "+
+		"miner fee: %v, prepay: %v", amount, quote.SwapFee,
+		quote.MinerFee, quote.PrepayAmount)
+
+	// Check that the estimated fees for the suggested swap are
+	// below the fee limits configured by the manager.
+	feeReason := m.checkFeeLimits(quote, amount)
+	if feeReason != ReasonNone {
+		return nil, newReasonError(feeReason)
+	}
+
+	outRequest, err := m.makeLoopOutRequest(
+		ctx, amount, balance, quote, autoloop,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return &outRequest, nil
 }
 
 // getSwapRestrictions queries the server for its latest swap size restrictions,
@@ -871,22 +1051,23 @@ func (m *Manager) getSwapRestrictions(ctx context.Context, swapType swap.Type) (
 // dispatched, and decides whether we set a sweep address (we don't bother for
 // non-auto requests, because the client api will set it anyway).
 func (m *Manager) makeLoopOutRequest(ctx context.Context,
-	suggestion *LoopOutRecommendation, quote *loop.LoopOutQuote,
+	amount btcutil.Amount, balance *balances, quote *loop.LoopOutQuote,
 	autoloop bool) (loop.OutRequest, error) {
 
 	prepayMaxFee := ppmToSat(
 		quote.PrepayAmount, m.params.MaximumPrepayRoutingFeePPM,
 	)
 
-	routeMaxFee := ppmToSat(
-		suggestion.Amount, m.params.MaximumRoutingFeePPM,
-	)
+	routeMaxFee := ppmToSat(amount, m.params.MaximumRoutingFeePPM)
+
+	var chanSet loopdb.ChannelSet
+	for _, channel := range balance.channels {
+		chanSet = append(chanSet, channel.ToUint64())
+	}
 
 	request := loop.OutRequest{
-		Amount: suggestion.Amount,
-		OutgoingChanSet: loopdb.ChannelSet{
-			suggestion.Channel.ToUint64(),
-		},
+		Amount:              amount,
+		OutgoingChanSet:     chanSet,
 		MaxPrepayRoutingFee: prepayMaxFee,
 		MaxSwapRoutingFee:   routeMaxFee,
 		MaxMinerFee:         m.params.MaximumMinerFee,
@@ -1095,31 +1276,33 @@ func newSwapTraffic() *swapTraffic {
 // maySwap returns a boolean that indicates whether we may perform a swap for a
 // peer and its set of channels.
 func (s *swapTraffic) maySwap(peer route.Vertex,
-	chanID lnwire.ShortChannelID) Reason {
+	channels []lnwire.ShortChannelID) error {
 
-	lastFail, recentFail := s.failedLoopOut[chanID]
-	if recentFail {
-		log.Debugf("Channel: %v not eligible for suggestions, was "+
-			"part of a failed swap at: %v", chanID, lastFail)
+	for _, chanID := range channels {
+		lastFail, recentFail := s.failedLoopOut[chanID]
+		if recentFail {
+			log.Debugf("Channel: %v not eligible for suggestions, was "+
+				"part of a failed swap at: %v", chanID, lastFail)
 
-		return ReasonFailureBackoff
-	}
+			return newReasonError(ReasonFailureBackoff)
+		}
 
-	if s.ongoingLoopOut[chanID] {
-		log.Debugf("Channel: %v not eligible for suggestions, "+
-			"ongoing loop out utilizing channel", chanID)
+		if s.ongoingLoopOut[chanID] {
+			log.Debugf("Channel: %v not eligible for suggestions, "+
+				"ongoing loop out utilizing channel", chanID)
 
-		return ReasonLoopOut
+			return newReasonError(ReasonLoopOut)
+		}
 	}
 
 	if s.ongoingLoopIn[peer] {
 		log.Debugf("Peer: %x not eligible for suggestions ongoing "+
 			"loop in utilizing peer", peer)
 
-		return ReasonLoopIn
+		return newReasonError(ReasonLoopIn)
 	}
 
-	return ReasonNone
+	return nil
 }
 
 // checkFeeLimits takes a set of fees for a swap and checks whether they exceed